pyproject.toml
[tool.poetry]
name = "d2r"
version = "0.1.0"
description = ""
authors = ["Andrey Raspopov <raspopov@arrival.com>"]

[tool.poetry.dependencies]
python = "^3.10"
six = "^1.16.0"
python-snappy = "^0.6.1"
numpy = "^1.23.4"
pygame = "^2.1.2"
opencv-python = "^4.6.0"
requests = "^2.28.1"
protobuf = "^4.21.9"

[tool.poetry.dev-dependencies]
pylint = "^2.15.5"
flake8 = "^5.0.4"
black = "^22.10.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
